持久层添加flyway 做数据库版本控制

+++ repository/src/main/java/com/ouzhx/repository/mapper/CityMapper.java
@@ -3,13 +3,13 @@ package com.ouzhx.repository.mapper;
 import com.ouzhx.entity.City;
 
 public interface CityMapper extends BaseMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(City record);
 
     int insertSelective(City record);
 
-    City selectByPrimaryKey(Integer id);
+    City selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(City record);
